@@ -16,31 +16,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.4",
-    "@angular/common": "~8.2.4",
-    "@angular/compiler": "~8.2.4",
-    "@angular/core": "~8.2.4",
-    "@angular/forms": "~8.2.4",
-    "@angular/platform-browser": "~8.2.4",
-    "@angular/platform-browser-dynamic": "~8.2.4",
-    "@angular/router": "~8.2.4",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.5.7",
-    "jquery": "^3.3.1",
-    "popper.js": "^1.14.7",
-    "rxjs": "~6.4.0",
+    "@angular/animations": "^9.1.1",
+    "@angular/common": "^9.1.1",
+    "@angular/compiler": "^9.1.1",
+    "@angular/core": "^9.1.1",
+    "@angular/forms": "^9.1.1",
+    "@angular/platform-browser": "^9.1.1",
+    "@angular/platform-browser-dynamic": "^9.1.1",
+    "@angular/router": "^9.1.1",
+    "bootstrap": "^4.4.1",
+    "core-js": "^2.6.11",
+    "jquery": "^3.4.1",
+    "popper.js": "^1.16.1",
+    "rxjs": "~6.5.5",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.2",
-    "@angular/cli": "~8.3.2",
-    "@angular/compiler-cli": "~8.2.4",
-    "@angular/language-service": "~8.2.4",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "~0.901.1",
+    "@angular/cli": "^9.1.1",
+    "@angular/compiler-cli": "^9.1.1",
+    "@angular/language-service": "^9.1.1",
     "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "electron": "1.8.8",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -48,10 +48,10 @@
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
+    "karma-jasmine-html-reporter": "^1.5.3",
+    "protractor": "^5.4.3",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "typescript": "~3.8.3"
   }
 }
